Refactor JS
