+++ app/javascript/packs/application.js
@@ -9,11 +9,8 @@ import * as ActiveStorage from "@rails/activestorage"
 import "channels"
 
 
-require("design/active.js")
-require("design/bootstrap.min.js")
-require("design/jquery-2.2.4.min.js")
-require("design/plugins.js")
-require("design/popper.min.js")
+import("active.js")
+import("plugins.js")
 
 
 Rails.start()
